Initial commit
created playRound function which is the back bone of the game. There is probably room to make it more efficient.
For the time being it is working properly.

Created game function but it is not working, since it requires a system to reset the values for playerSelection and computeSelection after each round.